orchestrator/pyproject.toml
[project]
name = "orchestrator"
version = "0.1.0"
description = "EdgeOrchestra orchestrator server"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "grpcio-reflection>=1.68.0",
    "protobuf>=5.29.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "redis[hiredis]>=5.2.0",
    "zeroconf>=0.136.0",
    "structlog>=24.4.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "alembic>=1.14.0",
    "httpx>=0.28.0",
    "numpy>=1.26.0",
    "coremltools>=8.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "lz4>=4.3.0",
    "prometheus-client>=0.21.0",
    "scikit-learn>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "fakeredis[lua]>=2.21.0",
    "aiosqlite>=0.20.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "mypy-protobuf>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/orchestrator"]
artifacts = ["src/orchestrator/dashboard/templates/**", "src/orchestrator/dashboard/static/**"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
